# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TAGOBTOB Judging & Tabulation System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: auth
    description: Judge and organizer sessions
  - name: health
    description: Health operations
  - name: judges
    description: Judge management
  - name: scores
    description: Score submission and retrieval
  - name: tabulation
    description: Tabulation and leaderboard

paths:
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get the current browser session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
    delete:
      operationId: deleteSession
      tags: [auth]
      summary: Sign out of the current session
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /auth/organizer-sessions:
    post:
      operationId: createOrganizerSession
      tags: [auth]
      summary: Sign in as an organizer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessCodeInput"
      responses:
        "201":
          description: Organizer session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Organizer access is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/judge-sessions:
    post:
      operationId: createJudgeSession
      tags: [auth]
      summary: Sign in as a judge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JudgeAccessCodeInput"
      responses:
        "201":
          description: Judge session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid judge credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /judges:
    get:
      operationId: listJudges
      tags: [judges]
      summary: List all judges
      responses:
        "200":
          description: List of judges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Judge"
    post:
      operationId: createJudge
      tags: [judges]
      summary: Create a new judge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JudgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JudgeRegistration"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Three-judge roster is already full or the judge already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Organizer authorization is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /judges/{id}:
    delete:
      operationId: deleteJudge
      tags: [judges]
      summary: Delete a judge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Organizer authorization is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /judges/{id}/access-code:
    post:
      operationId: resetJudgeAccessCode
      tags: [judges]
      summary: Generate a new judge access code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
            minimum: 1
            multipleOf: 1
      responses:
        "200":
          description: Access code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JudgeRegistration"
        "403":
          description: Organizer authorization is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Judge not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: List scores, optionally filtered
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [group, solo]
        - name: judgeId
          in: query
          required: false
          schema:
            type: number
        - name: schoolCode
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
    post:
      operationId: submitScore
      tags: [scores]
      summary: Submit a score for a school entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreInput"
      responses:
        "201":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Three-judge roster is incomplete or this judge already scored the entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Judge authentication is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scores/{id}:
    delete:
      operationId: deleteScore
      tags: [scores]
      summary: Delete a score entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Organizer authorization is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tabulation/group:
    get:
      operationId: getGroupTabulation
      tags: [tabulation]
      summary: Get group category tabulation and leaderboard
      responses:
        "200":
          description: Group tabulation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TabulationResult"

  /tabulation/solo:
    get:
      operationId: getSoloTabulation
      tags: [tabulation]
      summary: Get solo (Mother Majorette) category tabulation and leaderboard
      responses:
        "200":
          description: Solo tabulation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TabulationResult"

  /tabulation/summary:
    get:
      operationId: getTabulationSummary
      tags: [tabulation]
      summary: Get a quick summary of both categories for dashboard use
      responses:
        "200":
          description: Summary of both categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TabulationSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: number
      required:
        - success
        - id

    Judge:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        createdAt:
          type: string
        hasAccessCode:
          type: boolean
      required:
        - id
        - name
        - createdAt
        - hasAccessCode

    JudgeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required:
        - name

    JudgeRegistration:
      allOf:
        - $ref: "#/components/schemas/Judge"
        - type: object
          properties:
            accessCode:
              type: string
              description: Share this one-time displayed code privately with the judge.
          required:
            - accessCode

    AccessCodeInput:
      type: object
      properties:
        accessCode:
          type: string
          minLength: 8
          maxLength: 256
      required:
        - accessCode

    JudgeAccessCodeInput:
      allOf:
        - $ref: "#/components/schemas/AccessCodeInput"
        - type: object
          properties:
            judgeId:
              type: number
              minimum: 1
              multipleOf: 1
          required:
            - judgeId

    Session:
      type: object
      properties:
        role:
          type: string
          enum: [anonymous, judge, organizer]
        judgeId:
          type: ["number", "null"]
      required:
        - role
        - judgeId

    Score:
      type: object
      properties:
        id:
          type: number
        judgeId:
          type: number
        judgeName:
          type: string
        category:
          type: string
          enum: [group, solo]
        schoolCode:
          type: string
        schoolName:
          type: string
        entryNo:
          type: string
        rawCriterion1:
          type: number
        rawCriterion2:
          type: number
        rawCriterion3:
          type: number
        deductionCount:
          type: number
        weightedCriterion1:
          type: number
        weightedCriterion2:
          type: number
        weightedCriterion3:
          type: number
        deductionTotal:
          type: number
        totalScore:
          type: number
        createdAt:
          type: string
      required:
        - id
        - judgeId
        - judgeName
        - category
        - schoolCode
        - schoolName
        - entryNo
        - rawCriterion1
        - rawCriterion2
        - rawCriterion3
        - deductionCount
        - weightedCriterion1
        - weightedCriterion2
        - weightedCriterion3
        - deductionTotal
        - totalScore
        - createdAt

    ScoreInput:
      type: object
      properties:
        category:
          type: string
          enum: [group, solo]
        schoolCode:
          type: string
          enum: ["01", "02", "03", "04", "05"]
        rawCriterion1:
          type: number
          minimum: 0
          maximum: 50
        rawCriterion2:
          type: number
          minimum: 0
          maximum: 20
        rawCriterion3:
          type: number
          minimum: 0
          maximum: 30
        deductionCount:
          type: number
          minimum: 0
          multipleOf: 1
      required:
        - category
        - schoolCode
        - rawCriterion1
        - rawCriterion2
        - rawCriterion3
        - deductionCount

    TabulationEntry:
      type: object
      properties:
        schoolCode:
          type: string
        schoolName:
          type: string
        entryNo:
          type: string
        judgeCount:
          type: number
        isComplete:
          type: boolean
        missingJudgeCount:
          type: number
        avgCriterion1:
          type: number
        avgCriterion2:
          type: number
        avgCriterion3:
          type: number
        avgDeduction:
          type: number
        avgTotalScore:
          type: number
        rank:
          type: ["number", "null"]
        award:
          type: ["string", "null"]
        prizeAmount:
          type: ["string", "null"]
        tieBreakApplied:
          type: boolean
          description: Whether the official tie-break policy was used to order this entry
        tieBreakReason:
          type: ["string", "null"]
          description: The first tie-break criterion that separated entries with equal final averages
        judgeScores:
          type: array
          items:
            $ref: "#/components/schemas/Score"
      required:
        - schoolCode
        - schoolName
        - entryNo
        - judgeCount
        - isComplete
        - missingJudgeCount
        - avgCriterion1
        - avgCriterion2
        - avgCriterion3
        - avgDeduction
        - avgTotalScore
        - rank
        - award
        - prizeAmount
        - tieBreakApplied
        - tieBreakReason
        - judgeScores

    TieBreakPolicy:
      type: object
      properties:
        status:
          type: string
          enum: [approved]
        title:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: string
        approvedBy:
          type: string
      required:
        - status
        - title
        - description
        - steps
        - approvedBy

    TabulationResult:
      type: object
      properties:
        category:
          type: string
          enum: [group, solo]
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TabulationEntry"
        tieBreakPolicy:
          $ref: "#/components/schemas/TieBreakPolicy"
        totalJudges:
          type: number
        requiredJudgeCount:
          type: number
        totalScoresSubmitted:
          type: number
      required:
        - category
        - entries
        - tieBreakPolicy
        - totalJudges
        - requiredJudgeCount
        - totalScoresSubmitted

    TabulationSummary:
      type: object
      properties:
        group:
          $ref: "#/components/schemas/TabulationResult"
        solo:
          $ref: "#/components/schemas/TabulationResult"
      required:
        - group
        - solo