# lib/api-zod/src/generated/types/accessCodeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TAGOBTOB Judging & Tabulation System API
 * OpenAPI spec version: 0.1.0
 */

export interface AccessCodeInput {
  /**
     * @minLength 8
     * @maxLength 256
     */
  accessCode: string;
}